# Optimize String#downcase and String#upcase for single byte optimizabl…