…e case

This is a very simple change that leads to a pretty good performance boost.

Here's the benchmark:

```crystal
require "benchmark"

string = "hEllO WoRlD!." * 10

Benchmark.ips do |x|
  x.report("downcase") do
    string.downcase
  end
  x.report("upcase") do
    string.upcase
  end
end
```

Before:

```
downcase   7.73M (129.38ns) (± 1.92%)  144B/op        fastest
  upcase   7.68M (130.15ns) (± 2.50%)  144B/op   1.01× slower
```

After:

```
downcase  42.20M ( 23.69ns) (± 3.39%)  144B/op        fastest
  upcase  41.84M ( 23.90ns) (± 3.83%)  144B/op   1.01× slower
```