@straight-shoota 2022-08-15
`Char#downcase`/`#upcase` essentially does pretty much the same thing, just with a bit indirection through some method calls.

I suppose the performance gain comes from skipping the `options.none? || options.ascii?` test for every single byte, because the entire loop already is already guarded by that condition.